extract permission sync into action class
